замечания после review 1

+++ stats/analyzer/src/main/java/ru/practicum/analyzer/repository/projection/RecommendedEventProjection.java
@@ -4,5 +4,5 @@ public interface RecommendedEventProjection {
 
     Long getEventId();
 
-    Float getSimilarity();
+    Double getSimilarity();
 }
